update upload file endpoint. Create/update world, parse chests and items and store in DB

+++ server/src/models.py
@@ -1,5 +1,5 @@
 from typing import Optional, List
-from sqlalchemy import String, Integer, Float, ForeignKey, DateTime
+from sqlalchemy import String, Integer, BigInteger, Float, ForeignKey, DateTime
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 from datetime import datetime, timezone
 class Base(DeclarativeBase):
@@ -75,7 +75,7 @@ class World(Base):
 
     # Unique identifier for the world
     uid: Mapped[int] = mapped_column(
-        Integer,
+        BigInteger,
         nullable=False,
         unique=True,
         index=True
@@ -83,9 +83,9 @@ class World(Base):
 
     version: Mapped[int] = mapped_column(Integer, nullable=False) # Version of the world
     net_time: Mapped[float] = mapped_column(Float, nullable=False) # Network time of the world
-    modified_time: Mapped[int] = mapped_column(Integer, nullable=False) # Last modified time
+    modified_time: Mapped[int] = mapped_column(BigInteger, nullable=False) # Last modified time (Unix timestamp in milliseconds)
     name: Mapped[str] = mapped_column(String(100), nullable=False) # Name of the world
-    seed: Mapped[int] = mapped_column(Integer, nullable=False) # Seed of the world
+    seed: Mapped[int] = mapped_column(BigInteger, nullable=False) # Seed of the world
     seed_name: Mapped[str] = mapped_column(String(100), nullable=False) # Seed name of the world
 
     # Timestamps for record keeping
